+++ thermion_dart/native/src/c_api/TRenderManager.cpp
@@ -45,11 +45,6 @@ EMSCRIPTEN_KEEPALIVE void RenderManager_render(TRenderManager *tRenderer, uint64
     renderManager->render(frameTimeInNanos);
 }
 
-EMSCRIPTEN_KEEPALIVE void RenderManager_requestRender(TRenderManager *tRenderer) {
-    auto *renderManager = reinterpret_cast<RenderManager *>(tRenderer);
-    renderManager->requestRender();
-}
-
 EMSCRIPTEN_KEEPALIVE void RenderManager_setPaused(TRenderManager *tRenderer, bool paused) {
     auto *renderManager = reinterpret_cast<RenderManager *>(tRenderer);
     renderManager->setPaused(paused);